Reformats naming standards and moves axiosDefault to indx.js
