+++ src/App.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import Explore from './pages/Explore'
+import Explore from './views/Explore'
 import './app.css'
 
  const App = () => {
