+++ src/components/ExploreList.jsx
@@ -1,6 +1,6 @@
 import React, {useState, useEffect} from 'react'
 import { Item, Segment, Header, Message } from 'semantic-ui-react'
-import {getExploreArticles} from '../modules/dataCenter'
+import {getExploreArticles} from '../modules/articlesDataModule'
 
 const ExploreList = () => {
   const [articles, setArticles] = useState([])
@@ -33,7 +33,7 @@ const ExploreList = () => {
 
   return (
     <Segment textAlign="left">
-      <Header className="explore-list-header">Latest Stories</Header>
+      <Header className="explore-list-header" data-cy="explore-list-header">Latest Stories</Header>
       {errorMessage ? (
         <Message data-cy="error-message" header={errorMessage}/>
       ): (
