+++ src/index.js
@@ -3,6 +3,9 @@ import ReactDOM from 'react-dom';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
 import 'semantic-ui-css/semantic.min.css'
+import axios from 'axios'
+
+axios.defaults.baseURL = 'http://localhost:3000/api'
 
 ReactDOM.render(
   <App />,
