+++ src/modules/articlesDataModule.js
@@ -1,9 +1,7 @@
 import axios from 'axios'
 
-axios.defaults.baseURL = 'http://localhost:3000'
-
 const getExploreArticles = async () => {
-  let response = await axios.get('/api/articles?article_type=story')
+  let response = await axios.get('/articles?article_type=story')
   return response.data.articles
 }
